Look for "assert" tranformation in `iflaststmtl`
With this commit, we start to see the benefits of a separating tree
transformation from source string generation, (aside from the
conceptual, beneifit and increasing modularity).

`iflastlstmt` and `ifstmt` we want to consider together for purposes
of transforming "if .. raise" into "assert".

However in terms of producing source text each of these has separate
rules for where to look to get information in producing source text.

And even if they didn't combining both concerns - look for tranformation
and generate string would make this already complex routine more
complex.