jump_if_{false,true}, -> POP_JUMP_IF_{FALSE,TRUE}
We need to simplify and the boolean logic mess.
Right now, we don't need the "jump_if.. abstraction, and that
complicates understanding rules, and rule checking (I need to use ast[x][0]
vs. ast[x])

The abstraction we currently need is in "and" and "or" logic, and that
will happen down the line.