Limit scope of iflaststmt...
Even though the Earley parser allow for ambiguous grammars,
reductions can prevent other reductions from occuring.

Here, iflaststmtl was preventing a chained-compare reduction.

iflaststmtl is kind of weird anyway, it wants to treat the if as a final
statement a different from a statement in certain cases where control flow
by virtue of a return or except just doesn't fall through.

This commit adds a check for fall-thorugh-ness and flags the construct
as not an "iflastsmt".