Introduce _BACK on POP_JUMP_IF...
This allows us to isolate the grammar and reduction checking better.
For example, an jump to an the "else" of an "if else" will never be
backwards.

This causes a little havoc to the grammar rules, because we need to now
introduce loop variants of grammar rules. However this is beneficial
because reduction tests can be made on those grammar rules without
having to look below at how the rule is composed.

Also, it helps in looking over the parse tree generated because
looping jumps and constructs are more clearly marked as such.